backend/pyproject.toml
[project]
name = "trading-backend"
version = "0.1.0"
description = "FastAPI 后端"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109",
    "uvicorn>=0.27",
    "trading-core",
    "trading-connectors",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "websockets>=15.0.1",
]

[tool.uv.sources]
trading-core = { workspace = true }
trading-connectors = { workspace = true }
